pyproject.toml
[project]
name = "splitbillsporfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.135.2",
    "gunicorn>=25.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "uvicorn>=0.42.0",
]
